Add files via upload
"Museum" project is another management application. It's interacting with the user through a login page and on the main page there are multiple menus for different functions. The background is constantlly changing, different pictures with the museums types managed by the application are displayed on screen. With this application you can book tickets for clients, check or delete reservations, store datas in tabels in database and send email confirmation to the client.
Have fun !