Add files via upload
"Puzzle game" or "Slizzing puzzle" is a game wher you have to reconstruct a picture divided vertically and orizontally in equal parts. The game is cutomizable so you can choose the picture and set a kind of difficulty by selecting number of slicing and number of moves when mixing the puzzle. By pressing 'N' key you can view index possiotion of the pieces on the board so the game to become easier to solve. The best thing is that you can upload more pictiures, take a selfie and try to solve your mixed face :)))
Are you challanging enough to give it a try ?